Allow multiple ways to specify the position of a layer
For many layouts its way more convenient to specify distance to the
screen edge or position from the center coordinate of the source so this
allows specifying several position contraints:

* x, y: put top-left corner of layer at position
* left: put left edge of the source this distance away from the left
  screen edge. Negative numbers normalize for the aspect ratio
* right: put right edge of the source this distance away from the right
  screen edge. Negative numbers normalize for the aspect ratio
* top: put top edge of the source this distance away from the top
  screen edge. Negative numbers are abs()'ed
* bottom: put top edge of the source this distance away from the bottom
  screen edge. Negative numbers are abs()'ed
* cx, cy: put the center of the layer at position
* scale: The size of the layer in ratio to the screen size

Unspecified variables will be calculated from the specified constraints.
For example a position and an edge location is enough to calculate a
scale and three edge locations are enough to calculate the position and
scale.