Implement full resolution of YUYV decoding
This decodes perfectly if the source and target resolution are the same and the source is full screen. It still looks bad when the source is scaled due to depending on nearest neighbor interpolation on the texture so the texture data isn't corrupted. linear interpolation would need to be done in the shader to improve this.