Implement fetching a frame from any of the sources through the API
The /api/media/{source} API now supports fetching a frame of any of the sources and now returns
a JPEG encoded version of the frame. A second variant was added with GET /api/media/{source}/{png,jpeg} to request a png or jpeg encoded version of the frame. PNG encoding is significantly slower but does have the upside of being able to retrieve a pixel perfect imagesource.

It is still only possible to write frames to ImageSources with this API.